pyproject.toml
[project]
name = "voicegptoss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cerebras-cloud-sdk==1.46.0",
    "fastapi==0.116.1",
    "pyngrok==7.3.0",
    "python-dotenv==1.1.1",
    "uvicorn==0.35.0",
]
